feat(corehandler): Remove corehandler
