--- cg/store/api/core.py
@@ -1,28 +1,0 @@
-import logging
-
-from cg.store.crud.create import CreateHandler
-from cg.store.crud.delete import DeleteDataHandler
-from cg.store.crud.read import ReadHandler
-from cg.store.crud.update import UpdateHandler
-from cg.store.database import get_session
-
-LOG = logging.getLogger(__name__)
-
-
-class CoreHandler(
-    CreateHandler,
-    DeleteDataHandler,
-    ReadHandler,
-    UpdateHandler,
-):
-    """Aggregating class for the store api handlers."""
-
-    def __init__(self, session):
-        DeleteDataHandler(session)
-        ReadHandler(session)
-        UpdateHandler(session)
-
-
-class Store(CoreHandler):
-    def __init__(self):
-        self.session = get_session()
